job-agent-gemini/ADK_Local/requirements.txt
streamlit
google-adk
python-dotenv
litellm